ajout utilitaires de creation de fichier pour interagir avec le cache stinger
