+++ src/main/kotlin/fr/convergence/proddoc/util/WSUtils.kt
@@ -21,38 +21,6 @@ object WSUtils {
 
     private val LOG: Logger = LoggerFactory.getLogger(WSUtils::class.java)
 
-    /**
-     *    crée un fichier temporaire à partir d'un tableau de bytes
-     *    et retourne un objet fichier
-     */
-    fun creeFichierTempBinaire(fichier: ByteArrayInputStream): File {
-        try {
-            LOG.debug("Début création du fichier binaire temporaire")
-            val fichierTemp = createTempFile(suffix = ".pdf")
-            fichierTemp.writeBytes(fichier.readBytes())
-            LOG.debug("Fin création du fichier binaire temporaire")
-            return (fichierTemp)
-        } catch (e: java.lang.Exception) {
-            if (e is IOException) {
-                throw (IllegalStateException("Problème d'écriture sur disque", e))
-            } else {
-                throw e
-            }
-        }
-    }
-
-    /**
-     *  crée et retourne l'URL de récupération du Kbis à partir d'un identifiant du cache
-     *  l'URL de base est un paramètre "applicatif"...
-     */
-    fun creeURLKbisLocale(numGestion: String): String {
-
-        val baseURL = "http://127.0.0.1:8080/"
-        val pathURL = "kbis/pdfnumgestion/$numGestion"
-
-        return (baseURL + pathURL)
-    }
-
     private fun creerURLApplimyGreffe() :String {
 
         // val baseURL = "http://172.31.4.97:8880"
